feat: add heatmap year navigation buttons - doc: update README.md - chore: split large JavaScript file into modular components for better maintainability
