+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hexo-graph",
-  "version": "1.0.8",
+  "version": "1.0.9",
   "description": "hexo-graph，一个基于echarts，集成博客热力图，博客月份统计图，分类统计图，标签统计图的多元化插件，并支持明暗配色",
   "main": "index.js",
   "scripts": {
